test config
Created a config file that holds the test variables and moved the visualization testing to its own file. Will see if datarobot stays initialized
