+++ tests/Project/test_Project_initialization.py
@@ -1,5 +1,4 @@
 import pytest
-
 import rapa
 
 # test rapa.Project initialization
