+++ tests/conf.py
@@ -1,0 +1,10 @@
+project_name = 'TUTORIAL_breast_cancer'
+project_id = '62b5dc8249ed6b10669ab468'
+
+featurelist_prefix = 'TEST_0.0.9'
+featurelist_name = featurelist_prefix + ' (25)'
+featurelist_id = '62b5de67a92c8927b1fd710b'
+
+starred_model_id = '62b5e1aeddc5c75c4d91cf7a'
+
+best_AUC_model_id = '62b5e1aeddc5c75c4d91cf84'
